Recognise when calls can be to local functions and do not transform those.
With this fix we close #10. We should never transform what might be a local function. If the user want a specific function he can use a qualified name. Still need to handle warnings in such cases, though, and to disable warnings when annotating the expressions before I have the bound variables annotation, see #24.